Fix CutGTests strict-link failure by archiving ord singleton symbols (#9563)
The SWIG-generated dbStaTCL_wrap.cxx (in dbSta.a) references
ord::OpenRoad::openRoad(), ord::OpenRoad::getDbNetwork() and
ord::getOpenRoad(), but those symbols were defined only in objects
compiled directly into the openroad executable (OpenRoad.cc) and the
ord SWIG library (OpenRoad.i) -- never into a static archive. Lazy/
shared workstation links resolve them at runtime, but strict static
linking (Nix, -Wl,-z,defs) fails with an undefined reference.

Split the small, low-dependency singleton accessors into a new
OpenRoadSingleton.cc, built as a static ord_app library, and link it
from the openroad executable, the ord SWIG lib, and (PUBLIC) the dbSta
SWIG lib so consumers of dbSta.a can resolve the symbols.

ord_app depends only on dbSta_lib (a leaf), so no library dependency
cycle is introduced; the heavyweight OpenRoad.cc stays in the
executable.

Verified: the symbols now live in libord_app.a; a strict re-link of
CutGTests with -Wl,-z,defs and --whole-archive on dbSta.a succeeds with
0 undefined references and all 14 tests pass; the full openroad build
still succeeds with exactly one definition of each symbol.
